+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# Pending
+# Thu Mar 19 20:30:38 UTC 2020
 
 * Open Measurement - Move Loaded Event Prior to Start [#13880](https://github.com/Applifier/unity-ads-webview/pull/13880)
-* Dealy opening ad if previous ad is still closing [#13627](https://github.com/Applifier/unity-ads-webview/pull/13627)
+* Add PausableListenerAPI for ABT-1125 [#13627](https://github.com/Applifier/unity-ads-webview/pull/13627)
+* Clear Event Queue Properly in PausableListenerAPI [#13887](https://github.com/Applifier/unity-ads-webview/pull/13887)
 * Developer age gate, part 1 [#13619](https://github.com/Applifier/unity-ads-webview/pull/13619)
-
+* [Auction V6] Cleanup AuctionProtocol Use [#13852](https://github.com/Applifier/unity-ads-webview/pull/13852)
 
 # Tue Mar 17 20:26:08 UTC 2020